+++ change-log.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [2.0]
 
+### Busy
+
+- migrate sample website from asp.net MVC to ASP.NET Core MVC so that we can run the website and test it using kestrel via dotnet run. 
+  - notes here : https://docs.microsoft.com/en-us/aspnet/core/migration/mvc?view=aspnetcore-2.2
+
 ### changed
 
 - now supporting latest Chrome drivers `ChromeDriver 73.0.3683.68`